// api/generated/model/measurementFormDtoBodyPart.ts
/**
 * Generated by orval v8.2.0 🍺
 * Do not edit manually.
 * LgymApi.Api
 * OpenAPI spec version: 1.0
 */

export type MeasurementFormDtoBodyPart = typeof MeasurementFormDtoBodyPart[keyof typeof MeasurementFormDtoBodyPart];


export const MeasurementFormDtoBodyPart = {
  Unknown: 'Unknown',
  Chest: 'Chest',
  Back: 'Back',
  Shoulders: 'Shoulders',
  Biceps: 'Biceps',
  Triceps: 'Triceps',
  Forearms: 'Forearms',
  Abs: 'Abs',
  Quads: 'Quads',
  Hamstrings: 'Hamstrings',
  Calves: 'Calves',
  Glutes: 'Glutes',
  BodyWeight: 'BodyWeight',
  Neck: 'Neck',
  Waist: 'Waist',
  Hips: 'Hips',
  Thigh: 'Thigh',
  BodyFat: 'BodyFat',
  Bmi: 'Bmi',
} as const;